refactor(bumblebee): change folder structure
