+++ tools/bumblebee-deno/README.md
@@ -1,7 +1,7 @@
 https://discord.com/developers/docs/topics/gateway
 
 Plan:
-1. Refactor
+1. DONE - Refactor
 2. Implement error handling with sending errors to the admin channel
 3. Decide do we need to have ability to ping as in previous version
 4. Decide how to make health check, ideal bot should send messages to health check instance
