--- tools/bumblebee-deno/src/error/index.ts
@@ -1,8 +1,0 @@
-export class Exception extends Error {
-  code: number;
-
-  constructor(component: string, message = 'Internal Server Error', code = 500) {
-    super(`[${component}] ${message}`);
-    this.code = code;
-  }
-}
