+++ tools/bumblebee-deno/src/services/task-queue.ts
@@ -1,0 +1,14 @@
+import { Queue } from '../utils/queue.ts';
+
+export const taskQueue = new Queue<() => Promise<void>>();
+
+const processQueue = async () => {
+  const task = taskQueue.get();
+
+  if (task && task === taskQueue.current) {
+    await task();
+    taskQueue.remove();
+  }
+};
+
+setInterval(processQueue, 2000);
