--- tools/bumblebee-deno/src/telegram/enums.ts
@@ -1,5 +1,0 @@
-export enum RestApi {
-  SendMessage = 'sendMessage',
-  EditMessageText = 'editMessageText',
-  DeleteMessage = 'deleteMessage',
-}
