+++ tools/bumblebee-deno/src/utils/exception.ts
@@ -1,3 +1,12 @@
+export class Exception extends Error {
+  code: number;
+
+  constructor(component: string, message = 'Internal Server Error', code = 500) {
+    super(`[${component}] ${message}`);
+    this.code = code;
+  }
+}
+
 export const handleAsyncError = async (fn: () => Promise<void>) => {
   try {
     await fn();
